pyproject.toml
[project]
name = "fastapi-study-room"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.129.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.11.0",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]

[tool.uv.workspace]
members = [
    "app/fastapi_basic",
]
